extra: split concrete meters/reporters from core, torch_compat to examples; 0.6.0
