+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "registry-pattern"
-version = "0.5.0"
+version = "0.6.0"
 description = "A flexible registry pattern with DI container, Pydantic validation, and recursive config building"
 readme = "README.md"
 requires-python = ">=3.8"
@@ -55,7 +55,7 @@ dependencies = [
 yaml   = ["pyyaml>=6.0"]
 # OpenTelemetryReporter (spans + Histogram metrics)
 otel   = ["opentelemetry-api>=1.20"]
-# registry.experimental.torch_compat (markers, TorchProfilerMeter, TensorBoardReporter)
+# examples/torch_compat.py (markers, TorchProfilerMeter, TensorBoardReporter)
 torch  = ["torch>=2.0", "tensorboard>=2.0"]
 docs = [
     "sphinx>=7.0",
@@ -83,7 +83,7 @@ Documentation = "https://registry-pattern.readthedocs.io/"
 Issues = "https://github.com/adnanhd/registry-pattern/issues"
 
 [tool.setuptools]
-packages = ["registry", "registry.mixin", "registry.experimental"]
+packages = ["registry", "registry.mixin", "registry.extra", "registry.integrations"]
 
 [tool.setuptools.package-data]
 registry = ["py.typed"]
