+++ tests/test_torch_compat.py
@@ -1,6 +1,7 @@
-"""Unit tests for ``registry.experimental.torch_compat`` -- markers, meter, reporter.
+"""Unit tests for the ``examples/torch_compat.py`` example -- markers, meter, reporter.
 
-All tests skip cleanly when torch is not installed.
+The example module is importable because ``tests/conftest.py`` puts ``examples/``
+on ``sys.path``. All tests skip cleanly when torch is not installed.
 """
 
 from __future__ import annotations
@@ -21,7 +22,7 @@ from registry import (  # noqa: E402 -- after importorskip
     detach_meter,
     detach_reporter,
 )
-from registry.experimental.torch_compat import (  # noqa: E402
+from torch_compat import (  # noqa: E402 -- examples/ is on sys.path (see conftest)
     BoundTo,
     Checksum,
     Device,
